[FIX] web_enterprise: fixed handle widget visiblity
Currently,
if x2many list view is in the tab, the first child of a row has given 32px of
padding-left in edit mode, and also the width of handle cell is 32px.
and because of this handle widget is not shown.

so, used the original padding-left for the handle cell in edit mode.

before this commit:
handle widget is not shown for an editable list view in edit mode.

after this commit:
handle widget will be shown for an editable list view in edit mode.

task- 2003535

closes odoo/enterprise#4559

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>