[FIX] mrp_mps: test failure on sunday
MPS configuration is per week and the default
delivery lead time on supplier is 1. In the test
an action replenish based on lead time is execute and
expect an empty behavior however since the test run on
sunday and the delivery lead time is 1. You could launch
the delivery for monday. The test is fix with a lead time
to 0 for the supplier.

closes odoo/enterprise#4590

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>