[IMP] helpdesk: multi company for aftersales flow
This commit sets the company_id required on helpdesk team. Indeed, as from a ticket
we can now generate several sort of document (repair, returns, ....), we need a company
in which create them. Moreover, an helpdesk team should be part of a company and can not
be shared, even if its members are cross company.

With that, we now propagate the company from the ticket (related from team, so required
too) to created repair orders, pickings, refunds, ...

Task-1999686

closes odoo/enterprise#4779

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>