[IMP] helpdesk: allow user to close own tickets
The purpose of this commit is to allow final users to close helpdesk
tickets. This will ease the work of support agents and reduce their
pipeline.

- Add a new boolean item on helpdesk team settings to allow clients to
close tickets they created.
- Add a "view ticket" link at the bottom of the mail sent after
ticket creation redirecting to the ticket's portal page, in order
to show the portal page to the customer and allowing him to close
his ticket.
- Add a button opening a modal on ticket's portal page when setting is
activated and ticket not yet in "closed" stage. This modal contains
a button which permits to close the ticket.

Task-2009722 Closes #4612

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>