[IMP] industry_fsm: display rows for user assigned to a task recently
This commit extends the read_user_group method to add the users which
had a task assigned to in the past 30 days.
This allows to easily create task from within the gantt view and assign
it to users which were recently active.
Without the read group, only users that have a task
fitting the gantt span are shown.

related task 2000719