[FIX] industry_fsm: make stopping timer then cancel wizard NOT reset the timer
When staring the timer on a fsm task then stopping it, a dialog opens to
ask for validation of the time entry. If canceling the dialog, the
timer is still considered as stopped and the time information is lost.

This is because the timer is reset as soon as the dialog is open.

This commit moves the timer rest in the action triggered by validating
the time entry. Which means if the dialog is closed, the counter will
have kept running.

related task 2000719