[IMP] account_reports: add the ability to sort by column Task 2010811
We can now sort a report according to a column. To so, click on the column in the browser.
To activate the functionality, you need to set the class 'sortable' in the column header (_get_columns_name) and add a key 'no_format' on the columns you want to sort. The sorting will be the default python sorting for the type of the 'no_format' value. The title and section rows also need to have the 'no_format' key.

This feature has been released for the aged balance reports.

closes odoo/enterprise#4618

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>