[IMP] pos: use iot scale
This changes the way `ScaleScreenWidget` works, so that it uses an IoT scale proxy
instead of `ProxyDevice.scale_read` to retrieve the weight.

Since the scale driver has a `get_status` method and is registered as an
"old" driver, we can still get the scale and connection statuses from the `DeviceProxy`
and don't need any additionnal rpc.

Task: 1892412