[FIX] project_forecast: non working days in grid
Currently, if you try to add a forecast during a non working day, you
get a validation error in the forecast's grid view, because the
adjustement mechanism is unable to find a date in the working
schedule.

This commit fixes the issue by fallbacking on default values (start
and end of the day) in the case of non working days.

Task-1985510