[IMP] industry_fsm: buttons visiblity cleanup
Based on the options active (or not) on the FSM project, the buttons
visiblity does not always make sense and can prevent the user from
successfully finishing the flow.

The aim of this commit is to adapt the visibility of the buttons to
the various possible configurations of a Field Service project.

Task-2009567