[IMP] industry_fsm: rework fsm overlapping
Right now the count of tasks overlapping each others is done through a
non stored computed field. This could be problematic regarding
performance as each task generate a search on many other task and this
needs to be done on each tasks.

This commits use a sql query to compute this in one go, therefore
helping performance.

There can be a lot of tasks that potentially overlap another one, so to
ease conflicts solving as a user, there is now a link to the next
conflicting task inside that warning on the task form

related task 2000719