[REF] project_forecast: recurrent forecast wizard
We need a wizard in forecast to create a forecast and a recurrence at the same time, with
only one form view. To implement a such mecanism, as we were in a hurry, we decided to
copy/paste some fields and behavior (the revelant one) from the forecast model to the wizard.
This brings a maintenance problem: ducplicated code is difficult to keep consistent.

This commit proposes a new way to implement this wizard: making the wizard model inherit from
forecast model, with the following consequences and rationales:
- all fields of forecast will be on wizard: not a big deal as want to create a forecast,
therefore, it makes senses to display any field on the form view (even for customization).
- onchange behavior are the same for forecast and wizard: completely wanted, as we
duplicated code
- same for constraints

Task-2025396

closes odoo/enterprise#4632

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>