[IMP] industry_fsm_report: default worksheet logic
Currently, when creating a task from Project's kanban view, the
Worksheet Template field was not correctly set as the parent project's
because that field is absent from the quick task creation "form".
This behavior is corrected by moving the setting of that field from
default_get to a default method on the field itself.

Task-2009567