[FIX] 2 Weeks Resource Calendar: Compute the right total hours by week
Currently, when we switch to a 2 week calendar, the amount of hour/week
is the sum of both weeks, so this amount is falsy. The work time rate is
based on the number of hours/week so this rate is also falsy.

Now, if we are in a 2 week calendar, the number of hours by week is
the mean over 2 weeks.