[FIX] mrp_workorder: adapt context depending on the model
The action 'mrp_workorder_action_tablet' can be called from two
places :
-The workorder kanban view
-The workorder tablet view

The context of this action is filled with the key active_id by default.
The issue is, active_id is supposed to be a workcenter id. from the tablet view,
active_id is a workorder_id. This lead to a wrong read on mrp.workcenter table

This commit remove the context of the action definition and fill it properly
when calling the action

closes odoo/enterprise#4663

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>