[FIX] l10n_be_hr_payroll_account: don't replace the context
Replacing the context may lead to inprevisible behaviours.
In this specific case, when the test are not executed, the context
will only contains the last installed company in the allowed_company_ids
(l10n_be_hr_payroll.res_company_be) after the install of
l10n_be_hr_payroll_account.

When installing pos_restaurant, no pricelist is found when creating
the pos.config since there is no corresponding pricelist for this
demo company.

This was leadind to red runbot when test are not executed.

closes odoo/enterprise#4665

Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>