[FIX] industry_fsm: fix naming conflict in onchanges
Two onchange method on the product template have the same name. One is
defined in sale_timesheet (community) and the other in fsm. Since the
one in fsm was defined after and lacked a call to super, the first one
didn't get called.

This commit adds the super call.

related task 2000719