[FIX] l10n_be_hr_payroll_fleet: Compute atn and co2 fees only for BE
Car ATN and CO2 fees computation are using some `rule.parameter` specific
to Belgium.
When the user is not logged in a Belgian Company, an `ir.rule` prevents
him to read the `rule.parameter` and the computation fails.

Since car ATN and CO2 fees are Belgian concept, it's acceptable to
compute them only for Belgian companies.