[FIX] hr_payroll: Don't read rule.parameter from cache when allowed_company_ids changed
Reading `rule.parameter` is performed with an `@ormcache`d method
to fasten payslip computation.

The cache is reused when allowed_company_ids is changed but
it shouldn't.

With 2 companies A and B in two different countries:

1. (In Company A) Create a `rule.parameter` for country of Company A

3. (In Company B) Try to read the `rule.parameter` => AccessDenied

5. (In Company A) Read the `rule.parameter` => ok, the value is cached

7. (In Company B) Try to read the `rule.parameter` => the value from cache is retrieve,
no AccessDenied is raised.

=> Add the allowed_company_ids in the cache key.

Sadly, `ormcache_context` cannot be used because `allowed_company_ids`
is a list which is unhashable. Hence it cannot be used for the cache key.
For that reason, the list is casted to a tuple.