[IMP] l10n_mx_reports: Taxes in group must taken into account in DIOT
Main
-

Let us say that you have One Grouped Taxed and that it is defined in
supplier. Children taxes in that one are to be reported to DIOT report
but as they are inside the definition of a grouped taxed the won't taken
into account though they are properly defined.

Conclusion
-

Children from grouped taxes must be taken into account to be fetched in DIOT
report

X-original-commit: ea04ce74fe696745201e6349a709111f8c1c2e9f