[FIX] hr_payroll: query counts
There are a certain number of extensions of CRUD operations in `hr` that
add some business logic.  Those operations behave slightly differently
depending on the current user, and using the superuser de facto discards
some of those extensions.  This is supposedly because the data found is
partly related to the real user, and partly related to the superuser.
Those shortcuts do not seem intentional in the code, but change the
behavior in surprising ways.