sale_subscription: correct revenue dates
PR #3296 refactored the deferred revenue models and somewhat tried
to adapt the sale_subscription code to replace the integration between
deferred revenues and subscriptions.

This adaptation did *not* go well.

The point of the integration was twofold:
1/ Allow havng dates on invoice lines reflecting the period of revenue
they impact (this is useful for the subscription dashboard)
2/ Allow to use deferred revenues for real on subscriptions' invoices,
which makes it possible to deffer your subscription revenues (duh)

The second feature is dropped entirely AFAICT, the first was somewhat
reimplemented but not well. The date fields for revenues were set as
related on the subscription (which makes no sense, as the dates of
revenues are only computed at invoicing and should certainly not
be changed upon update of the subscription) and the MRR was copied
from the subscription directly - while normally, the MRR should be
for the line itself and not a mirror of the whole subscription.

This commit shoehorns a better behaviour into these fields (who
were thankfully stored!):
- start/end date of subscription revenues are based on the date of
next invoice and periodicity of the subscription for "normal invoices"
and renewals and on the date of next invoice and the 'discount date'
for upsells
- MRR is computed based on the amount of the line and the dates

While the initial implementation was perhaps a bit hasty, this has
the big advantage of making a complex configuration of deferred revenues
unnecessary to have dashboard data in sale_subscription_dashboard.

Migration note: when asset_start_date and asset_end_date are present and
set, these fields should be transferred to their matching fields in this
commit but only if subscription_id is set on the invoice line (otherwise
your dealing with 'usual' assets and not subscription ones).

closes odoo/enterprise#4787

Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>