[FIX] timesheet_grid: remove useless expand_group on timesheet
When installing timesheet_grid, we got an expanded group on task. The idea was
to see the task from the previous grid period, in order to ease the record of
timesheets.
Problem is that this group expand method is only call when the ORM use a normal
`read_group`. For performance issue, the `section="1"` were removed from grid
view, making the view only calling `read_grid`(which is not using group expand
mechanism).
Moreover, because of that group_expand on the task_id field, when grouping
timesheet on the list view, we get an inconsistent result (see opw-2005866).

In conclusion, we got a method that is not doing what it was created for, and
messing up somewhere else. Removing it seems the better solution

Task-2025158

closes odoo/enterprise#5389

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>