[IMP] industry_fsm: add configuration and settings
This commit creates the new configuration menu for fsm.
Settings related to fsm were moved from the project settings to a new one dedicated to fsm.
Next to the main settings there is the project settings which display a list of all fsm projects, and allow the user to quickly edit the project's settings.
This also introduces a product settings view which display all products (template) that are related to a project in fsm mode.

This commit also adds a worksheet settings view to the fsm app, which allows to list create and edit worksheet templates for fsm reports.
It is accessed through the fsm's Configuration menu.

related task 2009563