[FIX] pos_iot: Payment terminal lines identif.
When using a payment terminal, all the updates that were received from
the IoT Box were applied to the selected payment line. The problem is
that you can select a payment line, then click on the "Send" button of
another one. The payment line that should receive the update is the
one to which the button belongs.

To avoid that, we add the `cid` of the concerned payment line to
transaction and reversal requests sent by the PoS. All the status
updates sent by the IoT Box include this ´cid´. We can then retrieve
the correct payment line.

In case of errors, we don't get the `cid` in the response of the Box.
The only way we can precisely identify the payment line which will be
in error is to find the one in status 'waiting' who is waiting for an
update.