[FIX] pos_iot: Lost payment confirmations
Sometimes, some updates coming from the IoT Box get lost and never
arrive in the frontend. Most of the time it is ok. In the case of
payment with payment terminals, this is critical. The confirmation of
a payment might get lost. The customer will be debited but the cashier
won't be able to validate the order.

To make sure that all updates are received, we query the Driver every 3
seconds once a transaction or reversal request has been sent. We stop
querying the driver if we have the confirmation that the payment is
completely done (`data.processing` is `False`)