[FIX] pos_iot: Payment cancel after screen change
When you started a transaction with a payment terminal, then changed
screens, the payment was still processing but the listener was removed.
The payment confirmations never reached the payment lines even though
the money was debited.

To avoid having lost payment confirmations, when changing screens, we
cancel the payment being processed.

This requires a few other changes:

    - We now create the listener when the payment is sent and remove it
    as soon as it is not being processed anymore, either due to a
    confirmation or an error. This means that we can send the 'Cancel'
    and be sure that the listener will be removed soon (either with the
    cancellation or with the confirmation if the 'Cancel' arrived too
    late) and that the status will be correct.

    - When selecting another order, we got the same problem but here,
    the current order has been changed when you get the confirmation of
    the cancellation. We then needed to add the possibility to manage
    updates destined to an order different than the active one.