[IMP] industry_fsm: FSM project configuration
Impacted modules: industry_fsm, industry_fsm_report, and
project_enterprise

This commit cleans and simplifies the set up of a FSM project. Before this
commit setting up an FSM project required to check many checkboxes (allow
timesheet, allow worksheet, ...) and complete every sub field required for
the feature to work. Moreover, we had to cover many cases and combinaison,
complexifying the flow of an FSM task. Some case did not make sense and it
was easy to have inconsistent configuration.

In order to simplify this, we introduced the 'is_fsm' flag. Checking it on a
project preconfigure the settings for the project to be in FSM mode. An FSM
project is a "task rate" billable project with timesheets and worksheets.
To do so, we
- introduce the 'is_fsm' flag, and the onchange forcing the configuration
- remove "allow planning": the gantt view and planned dates on task form view
are always visible. This is not an option anymore.
- remove "allow billable": we now use the flow of "task rate", meaning the
project can not be billable, but each task is at a different rate (the one
of the generated sale order line linked to the task, based on the service
product of the project).

As we allow end users to create multiple FSM project, it makes no sense to
restrict the tasks seen in FSM app to the one in the "Field Service" project.
We keep the project data, but open the filter of the actions in FSM app.

For non FSM project, the flags "allow timesheet" is still optional. The service
product is not relevant for those project. Normal project can use worksheet,
timesheet and timer. The project is flow is preserved as we don't see any FSM
feature on it (no "create invoice", no sending worksheet, ...).

Special thanks to MaxLarue for starting this change in its branch ;)

Task-2009567