[FIX] delivery_dhl: weight computation and representation
In the xsd file representing the xml structure we have to send to dhl
for requests, the type of 'Weight' type is a decimal having 3 digits. So
as the zeep library doesn't modify the representation of a python float
to match the number of digits needed by the xml. We are using
'float_repr' to only send 3 digits.

We've also converted the weight in the DCT request, as it was not taking
into account the UOM set in the main configuration.

closes odoo/enterprise#4738

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>