[MOV-IMP] website_crm_score : move lead identification via website to community
Since introduction of website_visitor, there is no need to
keep the lead identification via lead_id cookie when contact_form is sent.

Each time contact form is sent, a new lead is created and linked to
the visitor_id. So all the leads are linked between them
to the same visitor_id.

Lang_id has been moved to crm module. Has it's interesting to store the lang
of the lead, for marketing or follow up reason.

Pages_viewed machanism has been totally deleted.
The pages viewed by the lead is in fact stored on the website_visitor.
But anyway, this pages_viewed cookie was never set and was then unusable.

This commit is linked to community commit from https://github.com/odoo/odoo/pull/34624

Task ID : 2028059
PR #4834