[FIX]: l10n_uk_reports(_hmrc): adapt to new tax model and merge modules
With the new tax model in saas-12.3, we can not
use the financial report we had before to get the
values for the vat return sent to hmrc.  It changed
to the tax report lines used by the generic tax report.
The structure of the _get_lines is also slightly different.

Also, the module is there twice in saas-12.3 for the moment.
We meant to merge l10n_uk_reports_hmrc into l10n_uk_reports.

closes odoo/enterprise#4888

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>