[FIX] hr_payroll_account: fix the tests for empty moves
This is an adaptation for odoo/odoo#34786

The test test_00_hr_payslip was creating a account.move, but without
account.move.line in it because there were no structure and therefore no
rules associated to the hr.payslip. At least one of the rules has to
have account_credit or account_debit set.

Also, in action_payslip_done, don't create the move if it is empty.
While I was at it, I also transformed the create to a create_multi.

closes odoo/enterprise#4798

Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>