[FIX] web_studio: Allow to set groups
In studio, when adding an access group to a field,
the group immediately disapears from the many2many tag
widget. Also an empty list is sent to the server and the group
is therefore never set.

The cause is a concurrency issue. The method `_onFieldChanged`
registers the new value in the model and returns a promise that
resolves when it's done.
Here a `view_change` event was dispatched without waiting for the
model to be updated. Hence, the value was not set.
The promise returned by the super call needs to be awaited before
doing anything else. And it needs to be returned.

closes odoo/enterprise#4840

Signed-off-by: Martin Geubelle (mge) <mge@openerp.com>