[FIX] project: fix project overview link in kanban
This commit fixes a traceback when clicking on "Overview" in Project
kanban view with Forecast installed.

Since project forecast actions have been made into xml actions, there
is a problem in the way the context is retrieved by a stat button
method in project overview, considering it as a dict when it is now a
string to parse.

opw-2037280  Closes #4855

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>