[IMP] delivery_dhl: select payor for duties
When an international dutiable shipment was created, the person who had
to paid the duties was the sender, as it was hardcoded in the code. It
was possible to change this behavior by setting a ir_config_parameter
that was added in a bug fix.

We've now implemented a clean feature to select the payor of duties,
when a sipping method is dutiable, you can select if it is the sender or
recepient that should pay the duties.

To implement this dhl, we simply need to specify if it's sender or
receipient that pays the taxes if the service is dutiable.

TASK-ID: 1970027