[ADD] social_push_notifications: add social implementation for Push Notifications
This commit introduces the Web Push Notification implementation for the social app.

It is a little different from the others since it does not publish anything on an actual social media
but rather directly posts messages to actual website visitors through web push notifications.

Once the module is correctly configured, website visitors will be prompted by a popup asking them if they
want to "stay up to date with ..." and then ask them to turn on browser "Notifications".
If they do, we install a background JS worker on their system that will listen to incoming post to display
and show a notification when he receives one.
That is intended to work even if the user is not on the website anymore (whole purpose of the background worker).

! Please note that notification permission request requires an 'https' environment !

Social Administrators or Users can then create posts that are targeted on a specific audience based on
website.visitors properties (country, visited pages, ...).
The push notifications can be sent in the visitor's timezone based on a post parameter.

Posters will be able to see a preview of the post and even test the notification on their browser.

-------

Advanced usage:

If you want to use your own Firebase account, instead of built in IAP usage that requires an
enterprise subscription:

The module will intensively use the Firebase APIs, both on backend (firebase-admin python module)
and on frontend (JS framework) sides.
Please note that the firebase-admin python module has to be installed manually as it is not part
of the Odoo 'requirements.txt' (since it's not available in debian packages).

The user first has to create a Firebase application ( https://firebase.google.com/ ) and then configure
the Social Push Notification media in Odoo to be able to use push notifications.

-------

Task#2002000
Part of global social task #1936643

Co-authored-by: David Beguin <dbe@odoo.com>