[IMP] iot: Set keyboard layout in Iot Device form
We want to be able to set the keyboard layout in the form of the
IoT Devices. To do so, the following steps are needed:

- Get the list of possible layouts and variants. When the record for
the IoT Box is first created, we need to call a route defined on the
IoT Box to get the list of available layouts.
- Since the number of variants is quite important, we wanted to add
the possibility to search in the list. To do so, we needed to make
the list a Many2one and therefore create a model for the layouts.
- When saving the form, an action should be called on the IoT Box to
update the selected layout for the current device. This call needs to
done from the browser so we override the method that is called when
clicking on "Save" in the form.

TaskID: 1961025