[IMP] industry_fsm_report: disable worksheet edition in settings
When editing a worksheet template through studio in the settings, we actually land on a form view of a worksheet template. This worksheet template is actually temporary and not meant to be edited/created through this screen as worksheets need a related task which in the context of settings doesn't make sense.

Given the worksheet template is a dynamic model created on the fly, so is the related act_window action. This commit modify the generated action to disable create and edit actions. This is therefore applied everywhere which is probably better.

Right now the only place where we need worksheets to be editable is when coming from a task. However since this comes from a python action which overwrite the context it is possible to edit/save and do everything normally when going through that flow.

related task 2009563