[FIX] stock_barcode: fix tests
The mock functions used to patch the method write of stock.picking were using the attribute write of an instance of stock.picking as callback instead of the class method.
It prevented to override the method write of stock.picking as it created an infinite recursion loop between the override and the mock function.

task-19575