[IMP] point_of_sale: create one account move by session
Main goal of this task is to create single accounting entry
(AE) when closing a pos.session instead of individual AE for
each order in the session.

This significantly minimizes the number of created journal
entries by point_of_sale which also results to faster closing
of session when there is large numbers (order of thousands)
of orders. It also eases the reconciliation of payments because
on one hand, cash payments are automatically reconciled and on
the other, other payments are combine in a receivable line
based on the payment method.

TASK-ID: 1862388

closes odoo/enterprise#5068

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>