[IMP] web_studio, *: enable string/label edition using report editor
*:
	industry_fsm_report
	l10n_be_hr_payroll
	l10n_ca_check_printing
	l10_fr_hr_payroll
	l10n_mx_edi

In the current studio report editor, few of the labels / strings are
not editable (mostly the ones that are wrapped within th/td tags).

This commit removes the limitation by making those labels editable
with help of studio.

For that, we added support to directly edit the labels within th
and td tags, and also wrapped some labels / strings within span
tag wherever needed.

Task-Id 2030430

closes odoo/enterprise#4745

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>