[IMP] project_enterprise: task gantt view improvements
- The year scale is not very useful for task that are expected to
last around a single
- Since when in month scale, the slot size is pretty small (there
are ~60 of them side by side horizontally), splitting that space
into two makes the tasks pills very small. By setting the precision
to full day, it is possile to read the beginning of the task name.
- On project's gantt unassigned tasks are considered as 'open
shifts', we moved that specific line (undefined assigned to) at
the top of the gantt, because it makes it easier to dispatch tasks
to users. A js-class is needed here, since the ordering is made in
the client, base on a alphabetic sort.

Task-2009563