[IMP] project_enterprise: move the use_timesheet_timer setting on projects
This commit moves the settings that allows to use the timesheet timer on the project itself, allowing more control. The settings is automatically true if allow planing is too. Otherwise it is configurable. A related field was added on the task to be able to easily determine whether or not the setting should be applied to that particular task.

related task 2000719