[IMP] industry_fsm*: use sale line instead of product map
Impacted modules: industry_fsm, industry_fsm_report

This commit replaces the product maps by using the sale lines directly instead.
The rationnal comes from we want the same behavior as the sale order line: unit
price, handle discount, quantity, ... Having an intermediate model (the product
task map) complexify that as we have to maintain the data between sale lines
and map entries.

The primary goal was to see what was order (from the map) on the worksheet
report, before the SO was created. Instead of copying the sale line behavior,
the SO is now created in 'draft' when clicking on adding material products, or
when we want to see the worksheet. Timesheet can be created when not linked to
the SO; the existing one will be on sales order creation and the future one
will be linked directly (as 'task rate' is working in sale_timesheet).

Task-2009558