[IMP] web_enterprise: show avatar correctly when rectangle image
Before this commit,
The avatar shown in hamburger menu in mobile, doesn't correctly
look like a circle if the image used for avatar is rectangle.

After this commit,
The avatar shown in hamburger menu in mobile will be in circle
shape, irrelevant of the image is square or not.

task-2042276

closes odoo/enterprise#4971

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>