[FIX] l10n_mx_edi: Revert fields removed to payment
On this PR https://github.com/odoo/enterprise/pull/3944 was removed the
fields for payment method and bank account in the payment wizard, but
this fields are required to generate the payment complement, then, was
returned.

When the fields were removed not was removed the assignation in the
methods, then I get a traceback when try pay many invoices.

closes odoo/enterprise#5328

Signed-off-by: Josse Colpaert <jco@openerp.com>