[FIX] sign: re-allow uploading encrypted files
This was added in 12.0 to avoid customers trying to make it work
but actually developped for master. This user error should not have
been forward-ported to master at the time.

I must admit it's quite magical to see something like this work
from the start just by removing 2 lines of code :D