[IMP] iot, pos_iot: Open & Close terminal shift
Six payment terminals requires that a new shift is open everytime the
cashier changes.

We open a shift when the POS is started and close it when the POS is
closed or when the cashier changes.

We need to create listeners to retrieve the tickets that are generated
when a shift is open or closed. These listeners needs to be killed as
soon as the operation is done because another listener will be created
for the same device when creating payment lines.