[IMP] test_l10n_be_hr_payroll: Adapt test to odoo/odoo#35446
Rule parameter tests are tagged as post install because
when creating a company, the required field
`extract_show_ocr_option_selection` has a column in database
(with a non null constraint) but the field is not yet loaded in
the registry. Therefore, the default value is not used when inserting
in the database and the contraint fails.
Hence, the test cannot be executed at module update.

closes odoo/enterprise#5012

Related: odoo/odoo#35446
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>