[FIX] account_invoice_extract: fix buttons and boxes not appearing
Before this commit, when switching in edit mode between invoices parsed
by the OCR, the buttons and boxes of the invoice would not appear.

The bug was due to the hook used to detect when a new invoice was being
rendered on the attachment preview. The new hook used is more reliable
as it is called everytime the attachment content is updated.