[IMP] planning: copy previous week
This commit tranforms the "Duplicate Period" into a "Copy Previous Week"
feature. When creating planning for employees, it is rare to generate it
month by month. Moreover, this is more complex to handle based on the previous
month, since the first day of work plus 1 month might be a sunday (or any non
working day). Rather than complixying a feature, we prefer restrict it to a
simple and correctly working case.

Task-2024216