[IMP] planning: publication and controllers
This commit rework the sending email and planning to employees.

1/ single shift
When creating a single shift, it can be sent to the employee (if this one has
a work email set). A dedicated template is used to send it with all informations.

2/ sending planning
On gantt view, a button is displayed to allow manager to send the planning. They can
choose if the planning should include or not the "open shifts".
Employees of this planning (the ones having at least one shift) will receive either
- a link to a frontend view (if no access to the system)
- a link to the calendar in the webclient (if employee has access to the system)
No attachment are generated.
In any case, employees are able to assign themselve to open shifts through the link
received by email.

Task-2024216

Co-authored-by: gla-odoo <gla@odoo.com>
Co-authored-by: sbl-odoo <sbl@odoo.com>