[IMP] planning: provide 'color picker' field widget
For usability purpose, we need a widget to allow select a color. As a
color field is an "interger" in odoo, a widget with color block looks more
appopriate to help user choosing the color he wants.
This commit provides a such widget. We keep it in this module for now, but
it will probably interest other application and will end up in web addon.

Task-2024216