[FIX] web_studio: properly load studio assets if required
Before this commit, whenever we start a page with app_creator in its
url, the following steps would take place:

1. web client is started
2. home menu is created and started (with Create New App link)
3. studio lazy assets are loaded
4. application goes into 'studio' mode

However, it could happen that someone (or more likely, a tour) would
click on the create new app link before the studio lazy assets are
loaded. This would result in an error, since the studio client action
for creating new app would not have been loaded yet.

This commit changes slightly the previous steps, and just makes sure
that if we are in a studio page, we load the studio assets BEFORE we
actually create the home menu.

closes odoo/enterprise#5062

Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>